Create Rich Text File version of the text file
From 'Rules of the Thumb in Establishing Remote Repositories.txt'
To 'Rules_of_the_Thumb_in_Establishing_RemoteRepositories.rtf'

RTF is basically a text-based file that is still universal
(BUT not as universal as text files)
RTF is not vendor-specific
therefore can be still opened by different RTF editing apps.
An advantage of RTF over text files
is that it has more formatting options.